Refactored to make all loading code separate, and separate utils module
