+++ xbout/utils.py
@@ -1,0 +1,38 @@
+import numpy as np
+
+
+def _set_attrs_on_all_vars(ds, key, attr_data):
+    ds.attrs[key] = attr_data
+    for da in ds.values():
+        da.attrs[key] = attr_data
+    return ds
+
+
+def _check_filetype(path):
+    if path.suffix == '.nc':
+        filetype = 'netcdf4'
+    elif path.suffix == '.h5netcdf':
+        filetype = 'h5netcdf'
+    else:
+        raise IOError("Do not know how to read file extension {}"
+                      .format(path.suffix))
+    return filetype
+
+
+def _separate_metadata(ds):
+    """
+    Extract the metadata (nxpe, myg etc.) from the Dataset.
+
+    Assumes that all scalar variables are metadata, not physical data!
+    """
+
+    # Find only the scalar variables
+    variables = list(ds.variables)
+    scalar_vars = [var for var in variables
+                   if not any(dim in ['t', 'x', 'y', 'z'] for dim in ds[var].dims)]
+
+    # Save metadata as a dictionary
+    metadata_vals = [np.asscalar(ds[var].values) for var in scalar_vars]
+    metadata = dict(zip(scalar_vars, metadata_vals))
+
+    return ds.drop(scalar_vars), metadata
